pyproject.toml
[project]
name = "screenshot-translator"
version = "0.1.0"
description = "Clipboard screenshot OCR + JP translation using llama.cpp Qwen3-VL"
authors = [{ name = "" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pillow>=10.4.0",
  "httpx>=0.27.2"
]

[tool.ruff]
line-length = 100

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
